data_api/requirements.txt
Flask~=2.2.2
pika~=1.3.1
requests~=2.28.1
python-dotenv~=0.21.0